backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
pandas==2.2.3
openpyxl==3.1.2
python-multipart
python-dotenv==1.0.1
aiofiles
aiomysql==0.1.1
sqlalchemy==2.0.22
alembic==1.11.1
apscheduler==3.10.1
pydantic==2.5.1
pymysql